Move clients from CLI to Shakedown. (#6474)
Summary:
Shakedown uses the Python CLI package to access the clients for several
services. This diff moves the clients over to `shakedown.clients`.

JIRA issues: MARATHON-8386
