+++ shakedown/dcos/package.py
@@ -1,7 +1,9 @@
 import json
 import time
 
-from dcos import (cosmos, errors, package, packagemanager, subcommand)
+from dcos import subcommand
+from shakedown import errors
+from shakedown.clients import cosmos, packagemanager, package
 from shakedown.dcos.service import *
 
 import shakedown
